1. reorg code directory 2. add sqlite database for blockchain storage 3. fix flask app for jsonify data in request response
