+++ crypto/elgamal_util.py
