+++ test/encryption_test.py
@@ -1,4 +1,4 @@
-import time_release_blockchain.crypto.elgamal as elgamal
+import crypto.elgamal as elgamal
 import unittest
 import sys
