+++ test_miner.txt
