--- time_release_blockchain/default_miner_config.json
@@ -1,5 +1,0 @@
-{
-  "MINER_ADDRESS": "uWiVRoaGGKjH/WUcDyumsv05g0Y/o2qa1so9vcBMhm1cKwVJlefQ5O45SBEjykJSjwv1NV/qB6I0dnHR+ciF2Q==",
-  "MINER_NODE_URL": "http://127.0.0.1:5000",
-  "PEER_NODES": []
-}
